added `Path::identify()`/`Path::isHeader2()` and deprecated flawed `Path::isCPP()`, `Path::isC()` and `Path::isHeader()` (#4681)
None of the `Path::is*()` function is working as expected or consistent
(detailed via additional unit tests and deprecation comments).
I tried to fix them but that would have made made cases with headers
quite awkward to implement so I opted to generate a new function. This
also optimized `Path::acceptFile()` a file.

I am not 100% happy with including `settings.h` but I didn't see a place
to move this to. If #3774 gets merged we might be able to move the
`enum` into `keywords.h` (that could also be renamed).

I kept the old function in case somebody else is using these (they are
in the API after all) and will remove them after the next release.