Add Original Name to types for Typedefs (#5968)
Trying to Fix the issue described here:
https://sourceforge.net/p/cppcheck/discussion/general/thread/5775ff1666/

For the small example purpose script it works. Actually it shows the
`valueType-originalTypeName`. But i am not 100% sure, what is done. I
printed the tok2 and tok3 and in my simple case both where the same. And
if i add either of both the originalName it also works. But for now i
couldn't deduct what exactly tok2 and tok3 are doing