reduced `Tokenizer::isC()` usage (#5724)
Each `Token` (should) be connected to a `TokenList`. `Tokenizer` just
encapsulates that so we have no need to check the `Tokenizer` but can
simply ask the `Token`.

Also if we have function calls we pass in a flag to tell it if it is
C/C++ we can get rid of that flag and simply ask the `Token`.