Fix #12505 (cli: Add option --check-version to pin cppcheck version) (#6121)
I chose to not add it in the help output. But I don't have a strong
opinion it can be added there also.

This option was added for the safety certification. It could be very bad
if a user runs one version of cppcheck and thinks that he runs another
version.