mitigated most `bugprone-unused-return-value` clang-tidy warnings in test code when we enable it for *all* functions (#6673)
We should not have any unchecked returns within the test code to make
sure everything is working as expected (there might be a case that
should also apply to the production code).

Unfortunately enabling all functions is abusing this check and it will
produce a lot of "false positives" (e.g. `erase()`/`insert()` of
containers or usage of stream insertion operators) so this cannot be
enabled by default. Since `llvm:Regex` does not support negative
lookahead we cannot configure the check to exclude functions.